MudTabs: (WIP) Fix repositioning issue when Centered
The issue: When `Centered` is set to `true`, an incorrect panel index is
grabbed instead. This is due to the `margin: auto` style; somehow, the
`X` or `Left` attribute returned from `getClientBoundingRect` Web API
does not take into account the margin, causing the tabs origin position
to be erroneous.

The solution: The get the width of the centered tab panels (enclosed in
a wrapper), and substract this width by the width of the the
`mud-tabs-toolbar-content`. Divide the result by two to get the actual
tabs origin position.