Tabs: Assign ActivePanel even if index already assigned
Previous `setter` implementation of the `ActivePanelIndex` property
will not update the `ActivePanel` property if the  `_activePanelIndex`
is equal to the newly assigned index.

This causes panels that are shifted after panel removal (thus having new
index, the same as `_activePanelIndex`, not being assigned as the
`ActivePanel`).

This commit removes that `_activePanelIndex == value` check and simply
proceed with assignments.