Tabs: remove panel before updating active panel index
Previous implementation update the active panel index (when the removed
panel is the active panel) before actually removing the panel.

This commit make it so that the target panel is removed first before
assigning new active panel index. This is because during the active
panel index assignment to `ActivePanelIndex`, the `MudTabPanel`
associated with the assigned index will also be assigned to
`ActivePanel`. Assigning the new active panel index before the actual
removal of the target panel wouldn't take the re-indexing of the panels'
indexes into account.